Refactor Entities, altered id to Long

+++ ticketing-system/src/main/java/com/learning/ticketing/config/ApplicationAuditAware.java
@@ -21,6 +21,6 @@ public class ApplicationAuditAware implements AuditorAware<Integer> {
 
         User userPrincipal = (User) authentication.getPrincipal();
 
-        return Optional.ofNullable(userPrincipal.getId());
+        return Optional.ofNullable(userPrincipal.getId().intValue());
     }
 }
